Moved other api calls back to apiCalls.ts and created apiCalls folder
