+++ components/tables/TableCell.tsx
@@ -4,8 +4,8 @@ import { User } from "@/types/types";
 import { ColumnType } from "./AdminTable";
 import { useSelector } from "react-redux";
 import { RootState } from "@/store/store";
-import {getLatestEvent} from "@/utils/eventApiCalls";
-import { backfillUserEvents, removeUserFromEvents } from "@/utils/adminApiCalls";
+import {getLatestEvent} from "@/utils/apiCalls/eventApiCalls";
+import { backfillUserEvents, removeUserFromEvents } from "@/utils/apiCalls/adminApiCalls";
 
 interface TableCellProps<TData> extends CellContext<TData, unknown> {
   column: Column<TData, unknown>;
