+++ pages/execAppView.tsx
@@ -1,4 +1,4 @@
-import { getAllAppsByTerm, getAllTerms } from "@/utils/adminApiCalls";
+import { getAllAppsByTerm, getAllTerms } from "@/utils/apiCalls/adminApiCalls";
 import {
   Check,
   Download,
