--- utils/apiCalls.ts
@@ -1,10 +1,0 @@
-import axios from "axios";
-import store from "../store/store";
-
-//Other
-export const sendContactEmail = async (values: Record<string, string>) => {
-  await axios.post("/api/other/contact", values);
-};
-
-
-
