+++ utils/apiCalls/eventApiCalls.ts
@@ -1,10 +1,5 @@
 import axios from "axios";
-import store from "../store/store";
-
-//Other
-export const sendSponsorEmail = async (values: Record<string, string>) => {
-  await axios.post("/api/other/sponsor", values);
-};
+import store from "../../store/store";
 
 //Events
 export const getEvents = async (
